Fixed a bug w/ the intellisense being generated. One line was not converting the lib reference.
