+++ @types/helper/methods/index.d.ts
@@ -1,0 +1,9 @@
+export * from "./addContentEditorWebPart";
+export * from "./addScriptEditorWebPart";
+export * from "./createContentType";
+export * from "./createDocSet";
+export * from "./hasPermissions";
+export * from "./parse";
+export * from "./request";
+export * from "./setContentTypeFields";
+export * from "./setGroupOwner";
